enable dagger to inject database and Daos
Also changed stationId's type from Int to Long, as the Station's id is a
Long value.

+++ app/src/main/java/com/github/pksokolowski/smogalert/database/AirQualityLog.kt
@@ -32,7 +32,7 @@ data class AirQualityLog(
         val indexRaw: Int,
 
         @ColumnInfo(name = "station_id")
-        val stationId: Int,
+        val stationId: Long,
 
         @ColumnInfo(name = "error_code")
         val errorCode: Int,
